fix: Correct New Feed Selection test
Using jQuery contents() method to capture feed contents returned an
object that could not be compared with Jasmine tests. Object comparison
returned a false positive, making the tests appear to pass.

Refactored to use jQuery text() method to capture text portion of feed
contents as a string that could be compared with Jasmine tests.

Tested result.